requirements.txt

streamlit
openai
python-docx
pdfplumber
pandas
openpyxl
extract-msg
